value function manipulations on DBlock
